Upgrade to support CentOS7.
Note that CentOS-7.0-1406-x86_64-Minimal.iso is a buildlogs version because
centos.org hasn't released it just yet. It is also a beta EPEL repo that is used.

Because centos now uses XFS as the default filesystem, which can not easily be
write mounted on OS X, a small ISO containing the kickstart file is mounted
instead of writing it directly to a modified boot image. This means it is
necessary to interact with the installer and add a boot option. For this
inteaction the Parallels Virtualization SDK is used.